refactor: reorganize packages into presentation/domain/data 3-layer structure
- presentation/controller/: ApiResponseController, HomeController (moved from root)
- presentation/data/: DTOs (already here)
- presentation/: ApiEndPoint (moved from constants/)
- domain/model/: DeviceInfo, SaveHead, UserAssem (moved from model/)
- domain/: DateTimeConst (moved from constants/)
- data/dao/: DeviceInfoDao, UidMappingDao (moved from root)
- data/client/: KakakuClient (moved from root)
- data/util/: StringEncoder (moved from root)

Update all import declarations and make package-private constructors public
where cross-layer access is required.
