+++ src/main/java/jp/developer/bbee/pcassem/domain/firestore/FirestoreService.kt
@@ -1,8 +1,8 @@
 package jp.developer.bbee.pcassem.domain.firestore
 
-import jp.developer.bbee.pcassem.DeviceInfoDao.SaveItem
-import jp.developer.bbee.pcassem.model.SaveHead
-import jp.developer.bbee.pcassem.model.UserAssem
+import jp.developer.bbee.pcassem.data.dao.DeviceInfoDao.SaveItem
+import jp.developer.bbee.pcassem.domain.model.SaveHead
+import jp.developer.bbee.pcassem.domain.model.UserAssem
 import java.util.Optional
 
 interface FirestoreService {
