+++ src/main/java/jp/developer/bbee/pcassem/domain/model/SaveHead.kt
@@ -1,13 +1,12 @@
-package jp.developer.bbee.pcassem.model
+package jp.developer.bbee.pcassem.domain.model
 
 import java.time.LocalDateTime
 
 @JvmRecord
-data class UserAssem(
-    val id: String,
-    val deviceid: String,
-    val device: String,
+data class SaveHead(
+    val saveid: String,
     val guestid: String,
+    val savename: String,
     val createddate: LocalDateTime,
     val lastupdate: LocalDateTime,
 )
