+++ src/main/java/jp/developer/bbee/pcassem/presentation/ApiEndPoint.kt
@@ -1,4 +1,4 @@
-package jp.developer.bbee.pcassem.constants
+package jp.developer.bbee.pcassem.presentation
 
 object ApiEndPoint {
     const val GET_DEVICE = "/devicelist"
